fix(desktop): route support access through hooks
